Move Suggested to separate component and add templates and suggested to recent page
